requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
requests>=2.31
tabulate=0.9.0